clean up warnings on gcc

+++ ch02/rand.c
@@ -8,7 +8,7 @@ main()
 {
 	int i;
 
-	srand((int)time(NULL));
+	srand(time(NULL));
 	for (i = 0; i < 10; ++i)
 		printf("%d\n", rand());
